+++ pydocusaurus/renderer/__init__.py
@@ -22,13 +22,21 @@ Some test.
 from pkgutil import extend_path
 
 from . import components
+from . import saver
 from . import page
 from . import package
 
 from .page import render_obj
 from .package import render_package_as_mdx
 
-__all__ = ("components", "page", "package", "render_obj", "render_package_as_mdx")
+__all__ = (
+    "components",
+    "saver",
+    "page",
+    "package",
+    "render_obj",
+    "render_package_as_mdx",
+)
 
 # Set this local module as the prefered one
 __path__ = extend_path(__path__, __name__)