+++ pydocusaurus/renderer/package.py
@@ -20,95 +20,20 @@ API documentation.
 """
 
 import os
-import abc
 from importlib.metadata import version, PackageNotFoundError
 from types import ModuleType
-
-from typing import Any
-
-import json
 
 from .. import core
 from ..core.attree import build_sidebar
 from ..core.walker import ParserAbstract, PackageWalker, fast_import
+from .saver import SaverAbstract, SaverDefault
 from .page import RendererPage
 
 __all__ = (
-    "SaverAbstract",
-    "SaverDefault",
     "URIHolder",
     "PackageRenderer",
     "render_package_as_mdx",
 )
-
-
-class SaverAbstract(abc.ABC):
-    """The abstract file saver."""
-
-    @abc.abstractmethod
-    def save_text(self, file_path: str | os.PathLike[str], data: str) -> None:
-        """Save the text file.
-
-        Arguments
-        ---------
-        file_path: `str | PathLike[str]`
-            The path to the output file.
-
-        data: `str`
-            The text data to be saved.
-        """
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def save_data(self, file_path: str | os.PathLike[str], data: Any) -> None:
-        """Save the structured data (such as json).
-
-        Arguments
-        ---------
-        file_path: `str | PathLike[str]`
-            The path to the output file.
-
-        data: `Any`
-            The structured data to be saved.
-        """
-        raise NotImplementedError
-
-
-class SaverDefault(SaverAbstract):
-    """The default implementation of the file saver. It can be overriden if special
-    customization is required."""
-
-    def save_text(self, file_path: str | os.PathLike[str], data: str) -> None:
-        """Save the text file.
-
-        Arguments
-        ---------
-        file_path: `str | PathLike[str]`
-            The path to the output file.
-
-        data: `str`
-            The text data to be saved.
-        """
-        dir_path = os.path.dirname(file_path)
-        os.makedirs(dir_path, exist_ok=True)
-        with open(file_path, "w", encoding="utf-8") as fobj:
-            fobj.write(data)
-
-    def save_data(self, file_path: str | os.PathLike[str], data: Any) -> None:
-        """Save the structured data (such as json).
-
-        Arguments
-        ---------
-        file_path: `str | PathLike[str]`
-            The path to the output file.
-
-        data: `Any`
-            The structured data to be saved.
-        """
-        dir_path = os.path.dirname(file_path)
-        os.makedirs(dir_path, exist_ok=True)
-        with open(file_path, "w", encoding="utf-8") as fobj:
-            json.dump(data, fobj, indent=2)
 
 
 class URIHolder: